+++ components/raftstore/src/store/fsm/mod.rs
@@ -13,8 +13,8 @@ pub use self::{
     apply::{
         check_sst_for_ingestion, create_apply_batch_system, Apply, ApplyBatchSystem, ApplyMetrics,
         ApplyRes, ApplyRouter, Builder as ApplyPollerBuilder, CatchUpLogs, ChangeObserver,
-        ChangePeer, ExecResult, GenSnapTask, Msg as ApplyTask, Notifier as ApplyNotifier,
-        ParsedEntry, Proposal, Registration, SwitchWitness, TaskRes as ApplyTaskRes,
+        ChangePeer, ExecResult, GenSnapTask, Msg as ApplyTask, Notifier as ApplyNotifier, Proposal,
+        Registration, SwitchWitness, TaskRes as ApplyTaskRes,
     },
     peer::{new_admin_request, DestroyPeerJob, PeerFsm, MAX_PROPOSAL_SIZE_RATIO},
     store::{
